+++ CHANGES.rst
@@ -5,5 +5,11 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
+- Add german translations.
+  [pcdummy]
+
+- Added the mail action based upon contactlist.
+  [pcdummy]
+
 - Initial mrbob template.
   [pcdummy]